tdf#157788 Move "Show navigation panel" back to Presentation settings
Instead of hiding it in the general options dialog.

As written in https://bugs.documentfoundation.org/show_bug.cgi?id=157788#c4
"The options in Slide Show Settings should all apply to the workstation";
the options which are saved per user/workstation and not per document, should not be moved
away from that dialog.
Instead the other options should also be saved per user/workstation.

Change-Id: I720c949f08877abb8ef8f94dbcfa6c85f349631d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158808
Tested-by: Jenkins
Reviewed-by: Samuel Mehrbrodt <samuel.mehrbrodt@allotropia.de>