gbuildtojson: announce every target
vs-ide-integration takes tens of minutes on my system; and most
of the time, it didn't output anything to the terminal. Every
target handled by gbuildtojson takes ~2 s; with more than 870
targets, it makes sense to show the progress to avoid making an
impression of a hang.

The "GBJ" abbreviation (for "gbuildtojson") was not used in the
output before.

Change-Id: I71ab3ae59d1034f47140aad0dab6ed170140acb3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158846
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>