tdf#158031 editeng SID_*_HYPERLINK: use AlsoCheckBeforeCursor
This fixes the GTK3 problem where only the first half of the link
successfully ran SID_EDIT_HYPERLINK against an editeng hyperlink.

Fixed for Writer and Calc editeng hyperlinks (i.e. in textboxes).

TODO: fix SID_COPY_HYPERLINK (which doesn't SelectFieldAtCursor)

The problem showed up when the menu option was added
because the mouse was over an unselected field,
but when the mouse moved to select the menu option,
the slot was not considered valid anymore
if the cursor had moved behind the field.

(The right-click to open the context menu puts the cursor
before or after the field.)

Now, by checking for a field-before-the-cursor during validation,
the field is found and the command is not prevented from running.

Once this is all in place, some older commits from Caolan
should be able to be reverted.

(Strangely, SID_OPEN_HYPERLINK and SID_REMOVE_HYPERLINK
seem to run before the slot is invalidated.
No idea why they are different.
For GEN, the popup usually either kept the pre-menu mouse position,
or else it always runs the command before the slot invalidates.)

Change-Id: If992b3fdc94d89162b60447458cabced2d5e3f19
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158856
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins