lok: notebookbar: don't recreate toolbars too often
commit 	004512d87ab70587d95a708116a5c2f4b870565c
notebookbar: allow to create multiple instances for online

Changed SfxNotebookBar::StateMethod so in LOK case it always
enter to the code block which is meant to be used for
creating or reloading notebookbar. Change that so we check
if notebookbar was initialized for given view already.

Change-Id: Icfa1e572769043509479b7b98d055120ac3d1f74
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158872
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>