tdf#158014 Skia adds filled white areas in .svg when exporting to PDF
So my strategy here is to assume that ProcessAndBlurAlphaMask
was doing the right thing before
    commit 81994cb2b8b32453a92bcb011830fcb884f22ff3
    Date:   Fri Apr 16 20:33:10 2021 +0200
    Convert internal vcl bitmap formats transparency->alpha (II)
but the subsequent naiving changing of its logic undermines it
because of some subtle interaction.

So take the brute force approach of reverting most of the code
to its prior state (i.e. working in the transparency domain),
and doing an Invert() before and after the original code.

This seems to fix all of the test files I have on hand
for this situation for both Skia and non-Skia cases.

Change-Id: If4c4d4c5351a4ec55897bed96b57d28eda88f5dd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158793
Tested-by: Jenkins
Reviewed-by: Patrick Luby <plubius@neooffice.org>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>