tdf#135236 gtk3 a11y: Restore AtkObject's focus_event
After

    commit 14e6a810e14e28ab82fe65d66e604ce562016845
    Date:   Mon Aug 15 13:14:08 2022 +0100

        fix "invalid class cast from 'OOoAtkObj' to 'GtkWidgetAccessible'"

, Orca with the gtk3 VCL plugin was no longer announcing focus
(and content) for the "Not in dictionary" text edit in Writer's
spelling dialog ("Tools" -> "Spelling") when moving the focus
there e.g. using the Tab key.

Make it work again by restoring the `focus_event` function
from `AtkObject`, similar to how

    commit f0827c392641b45647241b3a22e24a95e2b595e3
    Date:   Fri Aug 19 16:13:04 2022 +0100

        Resolves: tdf#150496 we want the AtkObject get_parent function

        GtkAccessible overrode it, but we expect the AtkObject implementation

did for the `get_parent` function to fix the announcement
of Writer paragraphs after the above-mentioned commit.

The other functions that `gtk_widget_accessible_class_init`
overrides [1] are already explicitly set to the wrapper
ones (s. the lines above the ones added in this commit), so
don't need additional changes.

[1] https://gitlab.gnome.org/GNOME/gtk/-/blob/c47425c62efb99be8cd2f0eea651dda76a99ba04/gtk/a11y/gtkwidgetaccessible.c#L542-557

Change-Id: Iae3b47f7fa7fe5ca90acb045a077cf0af52dd7fa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158990
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>