tdf#158031 Revert "tdf#140361 use the DisableEditHyperlink state
... as of menu launch time"

This code is now obsolete,
so revert 7.3 commit d9a5302ca8bef90914539c12adb6c81496c28504.
(This cleanly reverted with the relevant portions of
 the following code-refactor commits also reverted.
    commit 29489b33b435683021b72cb2bce27aba8cb7a430
    commit 68e797402692c5c8abf1b2c4374e12a8d2707d07
)

It was a fine idea, but overly complex, dependent on invalidations,
and hard to re-implement everywhere it would be needed.

Instead, just also check before cursor for a field.

Change-Id: Id570fc4e33611b217eae8b2a2a85aa58e23d9d1a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159013
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>