tdf#158031 editeng SID_OPEN_HYPERLINK: use AlsoCheckBeforeCursor
This patch depends on prior patches for this bug report.

This fixes nothing being opened if the mouse was right-clicked
over that second half of the hyperlink (since a "smart" positioning
set the cursor after the field instead of before it).

Note that for some reason, OPEN always seemed to work with the mouse,
even though other hyperlink menu options using the same code
failed to find the field! So this code change is mainly for consistency.

However, I WAS able to make it always fail,
by moving the mouse away once the popup menu arrived,
and then using the KEYBOARD to "Open Hyperlink".
In that case, it always failed (because the field
was not selected, and not under the mouse).

Change-Id: I8fcd0386de6cea8c6a937afb6f63061ed62d4cb1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159016
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>