tdf#158031 editeng: rename GetFieldAtCursor
because it doesn't really get the field at the cursor,
and certainly is not similar to SelectFieldAtCursor.

It first gets the field under the mouse.
Then, if there is a selection, it gets the selected field,
else it looks for the field on either side.

There were LOTS of places where it
probably had not been used properly.

Most of those are gone now, so it is easier
to rename the function.

Change-Id: I1a64af24092582cf865509d2a474080258edd76c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159022
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins