tdf#158031 editeng SID_HYPERLINK_GETLINK: use GetFieldAtSelection
This is just a clean-up commit to eliminate the pointless
check of a field under the mouse pointer.
It is up to the caller to make sure that the field is selected
before calling the GETLINK function.

GETLINK is primarily used to get the link from a field
for the insert/edit hyperlink dialog.

This function SHOULD NOT check before the cursor for the link,
or else it will mess up the insert hyperlink if the user
is inserting one field directly after another one.
So, my changes here are just following what already existed in sw.

Change-Id: Iaa7699d2e5f2fdf26996aede8b35bd4e18b21b98
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159021
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins