SC navigator: fix changing language of contenttree
Save the actual language, when ScContentTree is created, and use
that language to localise its text.

Note: Saving language in ScContentTree::ScContentTree caused
problems, so it is better to save the language when this
tree is filled with text the first time.

In case of LOK, there can be separate views, with different
languages, that means there can be 2+ ScContentTrees, e.g. one
has text in English and the other in German.

When a new item is created that is listed in navigator, then every
ScContentTree is updated at once... but that would use the global
language, not the one what was used when the actual ScContentTree
was created.

Change-Id: I2dedf293e0ad9fb8f3cdd1090e1e1707a9f6cfa4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159077
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>