android: Don't add "-editing" suffix to version/vendor
Don't append "-editing" to the version string, which
would then end up to be something like
"24.2.0.0.alpha0+/7763190f7ec2/The Document Foundation-editing"
when the build was configured with
`--with-vendor="The Document Foundation"` and would result
in the "About" dialog/widget saying
"This release was supplied by The Document Foundation-editing."
(s. `AboutDialogFragment#onCreateDialog`).

While it's useful to have the different build flavors
(with or without the experimental editing support), I see no need
to have the build config reflected in the version string.
(Whether experimental editing support was enabled during the
build can easily be seen by checking whether the "Experimental
Mode" option is available in the settings.)

Change-Id: I48ddb3a842b9db4fc1f435683ed9d8e8e85898ed
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159079
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>