sw floattable, insert UI: inherit fly width from selected table width
Inline table width defaults to the body frame width, so a hardcoded 2cm
default for the fly width is a bit poor.

Leave the normal fly insert case unchanged, but if an entire table is
selected, then change the default to the table width.

Check for the table selection like SwFEShell::NewFlyFrame() does it, and
determine if the entire table is selected like SwDoc::MakeFlyAndMove()
does it.

With this, a default table with a default frame keeps its width on frame
insert.

Change-Id: Iaf954395a4799222074acd83b5eae52ca75ae0ae
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159104
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins