sw floattable, insert UI: inherit fly width from selected table width
Inline table width defaults to the body frame width, so a hardcoded 2cm
default for the fly width is a bit poor.

Leave the normal fly insert case unchanged, but if an entire table is
selected, then change the default to the table width.

Check for the table selection like SwFEShell::NewFlyFrame() does it, and
determine if the entire table is selected like SwDoc::MakeFlyAndMove()
does it.

With this, a default table with a default frame keeps its width on frame
insert.

(cherry picked from commit 0ecb69d53864b582eb59533729ada01d85d383e6)

Change-Id: Iaf954395a4799222074acd83b5eae52ca75ae0ae
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159203
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>