gtk4 a11y: Handle the "level" object attribute
Add initial handling/mapping for object attributes.

In LibreOffice and Gtk 3/ATK/AT-SPI, object attributes
are currently key-value pairs, so arbitrary attribute
names and values can be set.

For Gtk 4, there's currently a discussion on how
AT-SPI object attributes should be handled, s. [1].

One potential option is for them to be handled as
`GtkAccessibleProperty`s.

In any case, there's already a `GTK_ACCESSIBLE_PROPERTY_LEVEL`
property that matches the "level" object attribute that
Writer sets for headings to specify what heading level
this is (s. `SwAccessibleParagraph::getExtendedAttributes`)
and that the gtk3 VCL plugin reports as an ATK/AT-SPI
object attribute with the same name and semantics for AT-SPI,
which is in line with the specification in the Core
Accessibility API Mappings 1.2 [2].

Map that LO object attribute to the above-mentioned
`GTK_ACCESSIBLE_PROPERTY_LEVEL`.

The property is currently not yet mapped to an AT-SPI
attribute in Gtk 4, but together with a corresponding
merge request [3], the object attribute can be seen
in Accerciser as expected.

While object properties in LO are currently generally
only generated/updated when they're queried
(via `XAccessibleExtendedAttributes::getExtendedAttributes`),
the `GtkAccessibleProperty` handling in Gtk 4
would require to explicitly call
`gtk_accessible_update_property` with the new value.

This may require further adjustments on LO side to
keep the properties up-to-date (e.g. adding something
like a new `AccessibleEventId::OBJECT_PROPERTY_CHANGED`
event and then implementing corresponding handling in
all places that provide object properties).
But that's something to look into later, also depending
on the outcome of the discussion in [1].
(As of now, the gtk4 VCL plugin doesn't handle any
a11y events, not even the existing ones.)

[1] https://gitlab.gnome.org/GNOME/gtk/-/issues/6196
[2] https://www.w3.org/TR/core-aam-1.2/#ariaLevelHeading
[3] https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/6549

Change-Id: I024afd7b527a20922e69156e1562dda783be2b49
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159216
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>