tdf#158169 take no actions in disposed incarnation
Using the described steps in the bug report it is possible
to have an already disposed incarnation of ChildrenManagerImpl
that gets called with notifyEvent. If exposed it is necessary
not to execute actions, but to inform the caller by triggering
a lang::DisposedException(). The SfxBaseModel::notifyEvent
that executes this will then remove the instance from it's list.

Change-Id: Ie39e37d6e55ea66f441e732b612774b18d7c3ca1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159306
Tested-by: Jenkins
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>
(cherry picked from commit 23e9a4cf1eec75cdd40a695ecae2af9b952178d3)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159242
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>