Update SlideShow on DrawModel changes
It may happen that DrawModel changes while a SlideShow is running,
e.g. from EditView/Scripts/UNO API. We can try to suppress these
as we already try with EditView in that mode. As an alternative
it is also possible to react to these changes in the SLideShow
in a useful manner.
This is the 2nd try, main change is to do all of this directly
in SlideShow, so no chechs/tests in common code are needed.

Change-Id: Iaa393d78af1fc40003c83d896508abe31cd790d5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159447
Tested-by: Jenkins
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>