tdf#153194 writerfilter: RTF import: testContSectionPageBreak last para
For testContSectionPageBreak, Word inserts an additional empty paragraph
at the end of the document that was missing in Writer.

Ensure markLastSectionGroup() is always called at the end of the
document so that the \par that is dispatched in m_bNeedPar case will be
inserted and not automatically removed immediately.

Also add a test for the same document without \sbknone, which has 4
pages instead of 2.

Change-Id: Ib3e4fbdb66df55941e4a487d4b249cd98fe42008
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159472
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>