gtk4 a11y: Map BLOCK_QUOTE role to new gtk equivalent
Map the role to the `GTK_ACCESSIBLE_ROLE_BLOCK_QUOTE`
equivalent newly added to Gtk 4 in commit [1]:

    commit bbb0150383e6aabdbead635f1751322530392b3a
    Author:     Michael Weghorn <m.weghorn@posteo.de>

        a11y: Add block quote role

With this and the gtk commit in place, Accerciser shows
the block quote role for Writer paragraphs using the
"Block Quotation" paragraph style as expected.

Also needs a bump of the gtk version to >= 4.13.4
on the gtk "main" branch after the 4.13.3 release for
the `GTK_CHECK_VERSION` check to pass and actually make
use of the new role, see MR [2].
Omit the fallback to `GTK_ACCESSIBLE_ROLE_PARAGRAPH`
for Gtk versions >= 4.13.1 but < 4.13.4,
since they are all development versions for the
same stable release 4.14.

[1] https://gitlab.gnome.org/GNOME/gtk/-/commit/bbb0150383e6aabdbead635f1751322530392b3a
[2] https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/6574

Change-Id: I2c17da13867a61f0e6819913583392962cfd44a2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158685
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins