tdf#158204 Midnightblue template correct color
Color of the upper and lower strip of the first master slide was
declared in "Accent 1" style. This style is overwritten when applying
the master page from the side panel.
Color is now defined as overloading of standard style
All inheritances from Accent 1 and Accent 2 were removed for master page
Midnightblue
Other updates:
- page number position, size and font was corrected
- Liberation font declaration was missing
- remove German name of a style
- replace html code &apos; by '

Change-Id: I2e0ef69d7117719a9829535db24c94048603d9b1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159519
Tested-by: Jenkins
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>
Reviewed-by: Stéphane Guillou <stephane.guillou@libreoffice.org>
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/160142