sw: fix layout crash in SwObjectFormatterTextFrame::DoFormatObj()
Since commit 15b886f460919ea3dce425a621dc017c2992a96b this is happening
in testForcepoint93, but only on WNT.

The crash is in SwObjectFormatterTextFrame::DoFormatObj() calling
SwObjectFormatterTextFrame::InvalidateFollowObjs() and the mrPageFrame
has been deleted; there is also the m_bAgain flag set in the SwLayAction
up the stack, but the SwObjectFormatter::mpLayAction is null so
DoFormatObj() cannot check this condition.

In the caller CalcContent(), get the SwLayAction from the SwViewShell
and pass it to the SwObjectFormatter, which appears to fix the problem.

Change-Id: Ib3dc34ce39b4aad723251271c24bdd7a6598ab1c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159518
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>