sw floattable, delete UI: add an uno command to unfloat frame from context menu
Word has an easy UI to turn floating tables into inline ones. Writer had
a similar button, but that had a few shortcomings:

1) It was only visible if a non-split frame was too large to fit the
   page.

2) It was a separate VCL widget, so invisible to LOK clients.

3) It only worked for frames which had a single table in them.

Researching the problem, it's interesting how deleting a frame always
deletes its content as well, but e.g. deleting a section just removes
the container but leaves the content in the body text.

Fix the problem by adding a new menu item in the context menu that
always allows converting the frame to inline content at the anchor
point. This can share a bit of code with the old unfloat button.

The undo/redo still needs fixing, in a follow-up change.

Change-Id: I8ce05c9f958b08cb599fd5d2a27e770182f28cc7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159550
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins